Clean up server code
